+++ validator/impl/validate-query.hpp
@@ -323,7 +323,7 @@ class ValidateQuery : public td::actor::Actor {
   bool check_mc_validator_info(bool update_mc_cc);
   bool check_utime_lt();
   bool prepare_out_msg_queue_size();
-  void got_out_queue_size(size_t i, td::Result<td::uint32> res);
+  void got_out_queue_size(size_t i, td::Result<td::uint64> res);
 
   bool fix_one_processed_upto(block::MsgProcessedUpto& proc, ton::ShardIdFull owner, bool allow_cur = false);
   bool fix_processed_upto(block::MsgProcessedUptoCollection& upto, bool allow_cur = false);
